refactor: additional changes to project structure
