+++ package-parser/package_parser/cli/__init__.py
@@ -1,0 +1,1 @@
+from ._cli import cli
