+++ package-parser/package_parser/cli/_cli.py
@@ -1,0 +1,156 @@
+import argparse
+import json
+from argparse import _SubParsersAction
+from pathlib import Path
+from typing import Any, Optional
+
+from package_parser.cli._run_annotations import run_all_command, _run_annotations
+from package_parser.cli._run_api import _run_api_command
+from package_parser.cli._run_usages import _run_usages_command
+from package_parser.commands.usages import find_usages
+from package_parser.commands.api import distribution, distribution_version, get_api
+from package_parser.commands.dependencies import get_dependencies
+from package_parser.utils import ensure_file_exists
+
+_API_COMMAND = "api"
+_USAGES_COMMAND = "usages"
+_IMPROVE_COMMAND = "improve"
+_ANNOTATIONS_COMMAND = "annotations"
+_ALL_COMMAND = "all"
+
+
+def cli() -> None:
+    args = _get_args()
+    if args.command == _API_COMMAND:
+        _run_api_command(args.package, args.src, args.out)
+
+    elif args.command == _USAGES_COMMAND:
+        _run_usages_command(args.package, args.client, args.tmp, args.out)
+
+    elif args.command == _ANNOTATIONS_COMMAND:
+        _run_annotations(args.api, args.usages, args.out)
+
+    elif args.command == _ALL_COMMAND:
+        run_all_command(args)
+
+
+def _get_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(description="Analyze Python code.")
+
+    # Commands
+    subparsers = parser.add_subparsers(dest="command")
+    _add_api_subparser(subparsers)
+    _add_usages_subparser(subparsers)
+    _add_annotations_subparser(subparsers)
+    _add_all_subparser(subparsers)
+
+    return parser.parse_args()
+
+
+def _add_api_subparser(subparsers: _SubParsersAction) -> None:
+    api_parser = subparsers.add_parser(_API_COMMAND, help="List the API of a package.")
+    api_parser.add_argument(
+        "-p",
+        "--package",
+        help="The name of the package.",
+        type=str,
+        required=True,
+    )
+    api_parser.add_argument(
+        "-s",
+        "--src",
+        help="Directory containing the Python code of the package. If this is omitted, we try to locate the package "
+             "with the given name in the current Python interpreter.",
+        type=Path,
+        required=False,
+        default=None,
+    )
+    api_parser.add_argument(
+        "-o", "--out", help="Output directory.", type=Path, required=True
+    )
+
+
+def _add_usages_subparser(subparsers: _SubParsersAction) -> None:
+    usages_parser = subparsers.add_parser(
+        _USAGES_COMMAND, help="Find usages of API elements."
+    )
+    usages_parser.add_argument(
+        "-p",
+        "--package",
+        help="The name of the package. It must be installed in the current interpreter.",
+        type=str,
+        required=True,
+    )
+    usages_parser.add_argument(
+        "-c",
+        "--client",
+        help="Directory containing Python code that uses the package.",
+        type=Path,
+        required=True,
+    )
+    usages_parser.add_argument(
+        "-t",
+        "--tmp",
+        help="Directory where temporary files can be stored (to save progress in case the program crashes).",
+        type=Path,
+        required=True,
+    )
+    usages_parser.add_argument(
+        "-o", "--out", help="Output directory.", type=Path, required=True
+    )
+
+
+def _add_annotations_subparser(subparsers):
+    generate_parser = subparsers.add_parser(
+        _ANNOTATIONS_COMMAND, help="Generate Annotations automatically."
+    )
+    generate_parser.add_argument(
+        "-a",
+        "--api",
+        help="File created by the 'api' command.",
+        type=Path,
+        required=True,
+    )
+    generate_parser.add_argument(
+        "-u",
+        "--usages",
+        help="File created by the 'usages' command that contains usage counts.",
+        type=Path,
+        required=True,
+    )
+    generate_parser.add_argument(
+        "-o", "--out", help="Output directory.", type=Path, required=True
+    )
+
+
+def _add_all_subparser(subparsers: _SubParsersAction) -> None:
+    all_parser = subparsers.add_parser(
+        _ALL_COMMAND,
+        help="Run api and usages command in parallel and then run annotations command.",
+    )
+    all_parser.add_argument(
+        "-p",
+        "--package",
+        help="The name of the package.",
+        type=str,
+        required=True,
+    )
+    all_parser.add_argument(
+        "-s",
+        "--src",
+        help="Directory containing the Python code of the package. If this is omitted, we try to locate the package "
+             "with the given name in the current Python interpreter.",
+        type=Path,
+        required=False,
+        default=None,
+    )
+    all_parser.add_argument(
+        "-c",
+        "--client",
+        help="Directory containing Python code that uses the package.",
+        type=Path,
+        required=True,
+    )
+    all_parser.add_argument(
+        "-o", "--out", help="Output directory.", type=Path, required=True
+    )
