+++ package-parser/package_parser/cli/_shared_constants.py
@@ -1,0 +1,2 @@
+_API_KEY = "api"
+_USAGES_KEY = "usages"
