+++ package-parser/package_parser/commands/annotations/__init__.py
@@ -1,0 +1,1 @@
+from ._generate_annotations import generate_annotations
