+++ package-parser/package_parser/commands/dependencies/_dependency_patterns.py
