+++ package-parser/tests/commands/get_dependencies/test_get_dependency.py
@@ -1,5 +1,5 @@
 import spacy
-from package_parser.commands.get_dependencies import (
+from package_parser.commands.dependencies import (
     DependencyExtractor,
     extract_action,
     extract_condition,
